♻️ 그래프 1차 수정
* ✨ Price의 PriceRangeGraph 컴포넌트 분리

* ♻️ Price 컴포넌트 분리

* 🚚 컴포넌트 분리에 따른 route 수정

* ✨ constants 추가
