+++ src/constants/calendar.ts
@@ -1,4 +1,19 @@
 export const DAY_OF_WEEK = ["일", "월", "화", "수", "목", "금", "토"];
 
+export enum DateTarget {
+  CHECK_IN,
+  CHECK_OUT,
+  BETWEEN,
+}
+
+export enum CalendarTypes {
+  CHECK_IN,
+  CHECK_OUT,
+  ALL_REMOVE,
+}
+
+export const CANVAS_HEIGHT = 100;
+export const CANVAS_WIDTH = 365;
+export const BUTTON_RADIUS = 12;
 export const CALENDAR_PAGE = 2;
 export const HALF_MOVE_POINT = 50;
